+++ verify/next/vite.config.tsserver.ts
@@ -13,7 +13,10 @@ const config = defineConfig({
   build: {
     assetsInlineLimit: 0, // 1024 * 1024 * 128,
     lib: {
-      entry: path.resolve(__dirname, './node_modules/@codingame/monaco-vscode-typescript-language-features-default-extension/index.js'),
+      entry: path.resolve(
+        import.meta.dirname,
+        './node_modules/@codingame/monaco-vscode-typescript-language-features-default-extension/index.js'
+      ),
       name: 'tsserver',
       fileName: () => 'tsserver.js',
       formats: ['es'],
@@ -26,7 +29,7 @@ const config = defineConfig({
         inlineDynamicImports: false
       }
     },
-    outDir: path.resolve(__dirname, 'bundle/tsserver'),
+    outDir: path.resolve(import.meta.dirname, 'bundle/tsserver'),
     emptyOutDir: true
   }
 });