+++ packages/client/CHANGELOG.md
@@ -2,9 +2,9 @@
 
 All notable changes to this npm module are documented in this file.
 
-## [11.0.0-next.2] - unrelease
+## [11.0.0-next.3] - unrelease
 
-- Updated all `@codingame/monaco-vscode` packages to `36.0.0`
+- Updated all `@codingame/monaco-vscode` packages to `36.2.7`
 - Updated `vscode-languageclient` and `vscode-languageserver` to `10.1.0`, `vscode-languageserver-protocol` to `3.18.2` and `vscode-jsonrpc` to `9.0.1`.
 - Dropped eslint and rely fully on oxlint.