+++ packages/examples/CHANGELOG.md
@@ -4,8 +4,8 @@ All notable changes to this npm module are documented in this file.
 
 ## unreleased
 
-- Updated all `@codingame/monaco-vscode` packages to `36.0.0`.
-- Updated to `monaco-languageclient@11.0.0-next.2` and `@typefox/monaco-editor-react@8.0.0-next.2`.
+- Updated all `@codingame/monaco-vscode` packages to `36.2.7`.
+- Updated to `monaco-languageclient@11.0.0-next.3` and `@typefox/monaco-editor-react@8.0.0-next.3`.
 - Updated `langium`to `4.3.0`.
 - Updated `vscode-languageclient` and `vscode-languageserver` to `10.2.0`, `vscode-languageserver-protocol` to `3.18.2` and `vscode-jsonrpc` to `9.0.1`.
 - Dropped eslint and rely fully on oxlint.